requirements.txt
Django==2.2.*
djangorestframework==3.9.*
python-magic==0.4.*
WeasyPrint==47
